==> analysis/scripts/01_prevalence_literature-review.Rmd <==
---
title: "Prevalence amongst literature"
author: "Rebecca Hoyd"
date: "5/17/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
```

# Load data

```{r}
lit.files <- list.files("~/Documents/repos/exogieo/data/derived", full.names = T)
lit.files <- lit.files[!grepl("spak", lit.files)]

lit.ls <- lapply(lit.files, function(x) read.csv(x, stringsAsFactors = F))

prev.TCC <- readRDS("/fs/ess/PAS1695/projects/exotic/data/drake-output/2022-03-08/tcc_prevalence-taxonomy.RDS")
prev.TCGA <- readRDS("/fs/ess/PAS1695/projects/exotic/data/drake-output/2022-03-08/tcga_prevalence-taxonomy.RDS")

TCC.meta <- readRDS("/fs/ess/PAS1695/projects/exotic/data/drake-output/2022-03-08/tcc.meta.linkage.RDS")
TCGA.meta <- readRDS("/fs/ess/PAS1695/projects/exotic/data/drake-output/2022-03-08/tcga.meta.RDS")
site.resolutions <- read.csv(
  "/fs/ess/PAS1695/exoticpipe/external-data/members-of-TCGA-groups.csv",
  stringsAsFactors = F, check.names = F) %>%
  dplyr::select(-n)
```

# Format

## Full dataset

```{r}
exoRAtowideprev <- function(data, taxlev){
  tmp <- data
  tmp$Taxa <- tmp[[taxlev]]
  tmp.wide<- tmp %>%
    filter(!is.na(Taxa)) %>%
    group_by(sample, Taxa) %>%
    summarise(ra = sum(prev, na.rm = T)) %>%
    mutate(ra = ifelse(ra > 0, 1, 0)) %>%
    spread(key = "Taxa", value = "ra")

  return(tmp.wide)
}

exoToDF <- function(taxalevels = c("domain", "kingdom", "phylum", "class", 
                                    "order", "family", "genus", "species"), 
                     data){
  w.ls <- lapply(taxalevels, function(x) exoRAtowideprev(data, x))
  w.df <- reduce(w.ls, function(x,y) left_join(x,y)) 
  return(w.df)
}
```

```{r}
prev.tcc.w <- exoToDF(data = prev.TCC)
prev.tcga.w <- exoToDF(data = prev.TCGA)

prev.res.tcc <- lapply(prev.tcc.w[,-1], mean) %>%
  unlist() %>%
  as.data.frame() %>%
  rownames_to_column(var = "Taxon") %>%
  rename('ORIEN.prev' = '.') 

prev.res.tcga <- lapply(prev.tcga.w[,-1], mean) %>%
  unlist() %>%
  as.data.frame() %>%
  rownames_to_column(var = "Taxon") %>%
  rename('TCGA.prev' = '.') 
```

```{r}
lit.df <- bind_rows(lit.ls)

combined.dat <- lit.df %>%
  left_join(prev.res.tcc) %>%
  left_join(prev.res.tcga) %>%
  # filter(!(is.na(ORIEN.prev) & is.na(TCGA.prev))) %>%
  pivot_longer(cols = c('Prevalence', 'ORIEN.prev', 'TCGA.prev'), 
               names_to = "prev.type",
               values_to = "prevalence")
```

## Fuso in CRC

```{r}
prev.tcga.crc <- prev.tcga.w %>%
  left_join(TCGA.meta) %>%
  filter(TCGA.code %in% c("COAD", "READ"))

tcc.joincodes <- TCC.meta %>%
  left_join(site.resolutions) %>%
  mutate(sample = LibraryID) %>%
  select(sample, TCGA.code)

prev.tcc.crc <- prev.tcc.w %>%
  left_join(tcc.joincodes) %>%
  filter(TCGA.code %in% c("COAD", "READ", "OtherCR"))

tcc.crc.prev <- mean(prev.tcc.crc$g__Fusobacterium)
tcga.crc.prev <- mean(prev.tcga.crc$g__Fusobacterium)
```

# Plot 

```{r}

fuso_only_dat <- prev.res.tcc %>%
  left_join(prev.res.tcga) %>%
  filter(Taxon == "g__Fusobacterium") %>%
  pivot_longer(cols = c("ORIEN.prev", "TCGA.prev"),
               names_to = "Ref", values_to = "prevalence") %>%
  mutate(Reference = gsub(".prev", "", Ref))

combined.dat %>%
  filter(Taxon == "g__Fusobacterium") %>%
  filter(prev.type == "Prevalence")  %>%
  mutate(Reference = gsub("et al.", "et al.\n", Reference),
         Reference = gsub("Clinical Colorectal Cancer", "ClinColCan", Reference),
         Reference = gsub("Genome Research", "GenRes", Reference)) %>%
  bind_rows(fuso_only_dat) %>%
  add_row(Reference = "ORIEN (CRC)", prevalence = tcc.crc.prev) %>%
  add_row(Reference = "TCGA (CRC)", prevalence = tcga.crc.prev) %>%
  ggplot(aes(x = prevalence, y = Reference)) +
  geom_col(width = .8) +
  labs(x = "", y = "") +
  theme_classic() +
  theme(text = element_text(size = 7))
ggsave("../figures/barplot_prevalence_lit-compare.pdf", height = 2.5, width = 2.1)
```

# Check other percentages

```{r}
prev.tcga.split <- prev.tcga.w %>%
  left_join(TCGA.meta) %>%
  mutate(crc.stat = ifelse(TCGA.code %in% c("COAD", "READ"), "CRC", "other"),
         source = "TCGA") %>%
  select(sample, crc.stat, source, g__Fusobacterium)


prev.tcc.split <- prev.tcc.w %>%
  left_join(tcc.joincodes) %>%
  mutate(crc.stat = ifelse(TCGA.code %in% c("COAD", "READ", "OtherCR"), 
                           "CRC", "other"),
         source = "ORIEN") %>%
  select(sample, crc.stat, source, g__Fusobacterium)


combine.splits <- bind_rows(prev.tcc.split, prev.tcga.split)
```

```{r}
combine.splits %>%
  group_by(source, crc.stat) %>%
  summarise(prevalence = mean(g__Fusobacterium))
```
