==> analysis/scripts/02_correlations_microbes-with-clinical.Rmd <==
---
title: "Correlate microbes with clinical variables"
author: "Rebecca Hoyd"
date: "2/16/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(broom)
library(rlist)
```

# Load data

```{r}
data.dir <- "/fs/ess/PAS1695/projects/exotic/data/drake-output/2022-03-08"
tcc.exo.ra <- readRDS(file.path(data.dir, "tcc.exora.taxonomy.RDS"))
tcga.exo.ra <- readRDS(file.path(data.dir, "tcga.exora.taxonomy.RDS"))

tcga.clin <- readRDS(file.path(data.dir, "tcga.clin.tum.RDS"))
tcc.clin <- read.csv("/fs/ess/PAS1695/projects/exotic/data/2020-02-19_clinical_aggregated.csv")

tcga.scores <- read.csv(file.path(data.dir, "tcga_tmesig1.csv"))
tcc.scores <- read.csv(file.path(data.dir, "tcc_tmesig1.csv"))

tcc.cancer.labs <- readRDS(file.path(data.dir, "normalization.inputs.RDS"))[[2]] %>%
  filter(grepl("^SL", sample)) %>%
  select(sample,TCGA.code)
```



# Functions

```{r}
exoRAtowide <- function(data, taxlev){
  tmp <- data
  tmp$Taxa <- tmp[[taxlev]]
  tmp.wide<- tmp %>%
    dplyr::filter(!is.na(Taxa)) %>%
    group_by(sample, Taxa) %>%
    summarise(ra = sum(exo.ra, na.rm = T)) %>%
    spread(key = "Taxa", value = "ra")
  
  tmp.wide[is.na(tmp.wide)] <- 0
  return(tmp.wide)
}

exoToDF <- function(taxalevels = c("domain", "kingdom", "phylum", "class", 
                                    "order", "family", "genus", "species"), 
                     data){
  w.ls <- lapply(taxalevels, function(x) exoRAtowide(data, x))
  w.df <- reduce(w.ls, function(x,y) left_join(x,y)) 
  return(w.df)
}
```

```{r}
correlate_mics_clin <- function(mics, scores, data, datalab){
  cor.res <- lapply(mics, function(m) lapply(scores, function(i)
    try(cor.test(data[[m]], data[[i]],
                 method = "spearman") %>%
          tidy() %>%
          mutate(score = i, microbe = m, datset = datalab))) %>%
      list.clean(., is.character))
  
  
  test <- lapply(cor.res, function(x) bind_rows(x))
  cor.df <- bind_rows(test)
  return(cor.df)
}
```

# Formatting

```{r wide microbe data}
tcc.w <- exoToDF(data = tcc.exo.ra)
tcga.w <- exoToDF(data = tcga.exo.ra)

microbes <- colnames(tcc.w)[-1]
```

```{r formatting clinical}
clin.vars <- c("age", "buffa.score", "MitoScore", "bmi")

tcc.clin.form <- tcc.clin %>%
  mutate(sample = RNA.SL.ID,
         age = Age.at.Collection,
         bmi = BMI.at.Collection) %>%
  select(sample, age, bmi) %>%
  left_join(tcc.scores) %>%
  left_join(tcc.cancer.labs)

tcga.clin.form <- tcga.clin %>%
  mutate(sample = file_id.expression) %>%
  left_join(tcga.scores) %>%
  mutate(age = age_at_index/365.25,
         sample = file_id.BAM) %>%
  select(sample, age, bmi, buffa.score, MitoScore, cancer)
```

```{r model inputs}
tcc.modin <- tcc.clin.form %>%
  inner_join(tcc.w)

tcga.modin <- tcga.clin.form %>%
  inner_join(tcga.w)
```

# Run correlations

```{r, warning=FALSE}
tcc.res <- correlate_mics_clin(microbes, clin.vars, tcc.modin, "ORIEN") %>%
  mutate(padj = p.adjust(p.value, method = "fdr"))
tcga.res <- correlate_mics_clin(microbes, clin.vars, tcga.modin, "TCGA") %>%
  mutate(padj = p.adjust(p.value, method = "fdr"))
```

```{r}
tcc.cancers <- unique(tcc.modin$TCGA.code)
tcga.cancers <- unique(tcga.modin$cancer)

tcc.cph.cancer <- lapply(tcc.cancers, function(y)
  correlate_mics_clin(microbes, clin.vars,
                      filter(tcc.modin,
                             TCGA.code == y),
                      y))

tcga.cph.cancer <- lapply(tcga.cancers, function(y) 
  try(correlate_mics_clin(microbes, clin.vars,
                      filter(tcga.modin,
                             cancer == y),
                      y)))
```

```{r format split cancer results}
tcc.cancer.form <- bind_rows(tcc.cph.cancer) %>%
  mutate(cancer = datset,
         datset = "TCC",
         padj = p.adjust(p.value, method = "fdr")) 
tcga.cancer.form <- bind_rows(tcga.cph.cancer) %>%
  mutate(cancer = datset,
         datset = "TCGA",
         padj = p.adjust(p.value, method = "fdr")) 
```


# Save

```{r}
write.csv(tcc.res, "../data/correlations_mic-clin_TCC.csv", row.names = F)
write.csv(tcga.res, "../data/correlations_mic-clin_TCGA.csv", row.names = F)

write.csv(tcc.cancer.form, "../data/correlations_mic-clin_TCC_split-cancer.csv",
          row.names = F)
write.csv(tcga.cancer.form, "../data/correlations_mic-clin_TCGA_split-cancer.csv",
          row.names = F)
```

# Randomize Age

```{r}

randomAge <- function(s, dat){
  set.seed(s)
  tmp <- sample(dat$age, replace = F)
  return(tmp)
}

set.seed(8675309)
seeds <- runif(100, 1000000,9999999)
snames <- paste0("seed", 1:100)

tcc.rcols <- lapply(seeds, function(x) randomAge(x,tcc.modin))
tcga.rcols <- lapply(seeds, function(x) randomAge(x,tcga.modin))

names(tcc.rcols) <- snames
tcc.modin.r <- bind_cols(tcc.modin, tcc.rcols)

names(tcga.rcols) <- snames
tcga.modin.r <- bind_cols(tcga.modin, tcga.rcols)

save(tcc.modin.r, tcga.modin.r, microbes, file = "/fs/ess/PAS1695/projects/exotic/data/corr-inputs_mic-age-random.rda")
```

```{r, eval = FALSE}
tcc.res.r <- correlate_mics_clin(microbes, snames, tcc.modin.r, "ORIEN") %>%
  mutate(padj = p.adjust(p.value, method = "fdr"))
tcga.res.r <- correlate_mics_clin(microbes, snames, tcga.modin.r, "TCGA") %>%
  mutate(padj = p.adjust(p.value, method = "fdr"))
```

```{r, eval = FALSE}
tcc.cancers <- unique(tcc.modin.r$TCGA.code)
tcga.cancers <- unique(tcga.modin.r$cancer)

tcc.cph.cancer.r <- lapply(tcc.cancers, function(y)
  correlate_mics_clin(microbes, snames,
                      filter(tcc.modin.r,
                             TCGA.code == y),
                      y))

tcga.cph.cancer.r <- lapply(tcga.cancers, function(y) 
  try(correlate_mics_clin(microbes, snames,
                      filter(tcga.modin.r,
                             cancer == y),
                      y)))
```

```{r format split cancer results, eval = FALSE}
tcc.cancer.form.r <- bind_rows(tcc.cph.cancer.r) %>%
  mutate(cancer = datset,
         datset = "TCC",
         padj = p.adjust(p.value, method = "fdr")) 
tcga.cancer.form.r <- bind_rows(tcga.cph.cancer.r) %>%
  mutate(cancer = datset,
         datset = "TCGA",
         padj = p.adjust(p.value, method = "fdr")) 
```


## Set up visualization

```{r, eval = FALSE}
tcc.res <- tcc.res.r %>%
  mutate(cancer = "All") %>%
  bind_rows(tcc.cancer.form.r)

tcga.res <- tcga.res.r %>%
  mutate(cancer = "All") %>%
  bind_rows(tcga.cancer.form.r)

corr.res <- bind_rows(tcc.res, tcga.res)

matched.cancers <- unique(subset(corr.res, corr.res$datset == "TCGA")$cancer)
```

```{r, eval = FALSE}
tcc.newlab <- tcc.res %>%
  filter(p.value < 0.05) %>%
  mutate(TCC.est = estimate, 
         dir = TCC.est >= 0) %>%
  select(TCC.est, dir, microbe, score, cancer)

tcga.newlab <- tcga.res %>%
  filter(p.value < 0.05) %>%
  mutate(TCGA.est = estimate, 
         dir = TCGA.est >= 0) %>%
  select(TCGA.est, dir, microbe, score, cancer)

corrs.matched.1 <- inner_join(tcc.newlab, tcga.newlab)

# corrs.matched <- tcc.newlab %>%
#   filter(cancer %in% c("LUAD", "LUSC", "SARC", "KIRC", "BLCA") & score == "bmi") %>%
#   bind_rows(corrs.matched.1)

corrs.matched.1 %>%
  group_by(cancer, score) %>%
  tally() %>%
  arrange(score)

saveRDS(corrs.matched.1, "../data/correlations_mic-clin_randomized.RDS")
```

```{r, eval = FALSE}
matched.long <- corrs.matched.1 %>%
  pivot_longer(cols = c(TCC.est, TCGA.est),
               names_to = "datset", values_to = "estimate")

matched.long %>%
  ggplot(aes(y = microbe,x = estimate,
             yend = microbe, xend = 0)) +
  facet_grid(cancer ~ score) +
  geom_segment() +
  geom_point(aes(color = datset)) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Correlation estimate", y = "") +
  scale_color_manual(breaks = c("TCC.est", "TCGA.est"),
                     labels = c("ORIEN", "TCGA"),
                     values = c("red", "blue"),
                     name = "",
                     ) +
  theme_classic() +
  theme(text = element_text(size = 10),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        legend.position = "top")
ggsave("../figures/lollipop_mic-clin_randomize.pdf", height = 6, width = 4)
```

# Randomize Buffa

```{r}
randomBuffa <- function(s, dat){
  set.seed(s)
  tmp <- sample(dat$buffa.score, replace = F)
  return(tmp)
}

set.seed(8675309)
seeds <- runif(100, 1000000,9999999)
snames <- paste0("seed", 1:100)

tcc.rcols <- lapply(seeds, function(x) randomBuffa(x,tcc.modin))
tcga.rcols <- lapply(seeds, function(x) randomBuffa(x,tcga.modin))

names(tcc.rcols) <- snames
tcc.modin.r <- bind_cols(tcc.modin, tcc.rcols)

names(tcga.rcols) <- snames
tcga.modin.r <- bind_cols(tcga.modin, tcga.rcols)

save(tcc.modin.r, tcga.modin.r, microbes, file = "/fs/ess/PAS1695/projects/exotic/data/corr-inputs_mic-buffa-random.rda")
```
