==> analysis/scripts/02_visualize_survival.Rmd <==
---
title: "visualize survival results"
author: "Rebecca Hoyd"
date: "2/23/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(ggforce)
library(devEMF)
library(ggtext)
```

# Load

```{r}
tcc.all <- readRDS("../data/survival_tcc_all-cancers.RDS") %>%
  bind_rows()
tcga.all <- readRDS("../data/survival_tcga_all-cancers.RDS") %>%
  bind_rows()
tcc.split <- readRDS("../data/survival_tcc_split-cancers.RDS")
tcga.split <- readRDS("../data/survival_tcga_split-cancers.RDS")
```

```{r}
tcc.all.join <- tcc.all %>%
  mutate(cancer = "All")
tcga.all.join <- tcga.all %>%
  mutate(cancer = "All")

surv.res <- bind_rows(tcc.all.join, tcga.all.join, tcc.split, tcga.split) %>%
  mutate(padj = p.adjust(p.value, method = "fdr"))

write.csv(surv.res, "../tables/S1_survival-results.csv", row.names = F)
```

```{r}
matched.cancers <- unique(subset(surv.res, surv.res$datset == "TCGA")$cancer)
```

# Circos

## Formatting

```{r make big plot input object}
tile.input <- surv.res %>%
  mutate(taxalev = str_sub(microbe,1,1),
         sigcode = p.value < 0.05) %>%
  select(microbe, taxalev, sigcode, datset, cancer)


```

```{r find the variables that we want to limit the plots on}
test <- tile.input %>%
  filter(sigcode == T & !grepl("Other", cancer)) %>%
  group_by(cancer) %>%
  tally()

good.cancers <- test$cancer
all.taxlevs <- unique(tile.input$taxalev)
```

## geom_tile

```{r ensure consistent fill scale}
vir.cols <- viridis::viridis_pal()(9)
names(vir.cols) <- good.cancers
```

```{r plotting function}
polarizedHeatmaps <- function(tlev, dset, yadj){
  tile.input %>%
    filter(datset == dset & taxalev == tlev) %>%
    filter(cancer %in% good.cancers) %>%
    mutate(cancer.code = ifelse(sigcode == T, cancer, NA),
           cancer = as.numeric(as.factor(cancer)) + yadj) %>%
    ggplot(aes(x = microbe, y = cancer, fill = cancer.code)) +
    geom_tile() +
    ylim(c(0,100)) +
    scale_fill_manual(values = vir.cols, name = "", na.value = "grey95") + 
    theme_void() +
    theme(axis.text.x = element_blank(),
          axis.ticks = element_blank()) +
    coord_polar()
}
```

## Plots for ORIEN and TCGA 

```{r}
ringplot.tcc.ls <- list(d = polarizedHeatmaps("d", "TCC", 1),
                        k = polarizedHeatmaps("k", "TCC", 12),
                        p = polarizedHeatmaps("p", "TCC", 23),
                        c = polarizedHeatmaps("c", "TCC", 34),
                        o = polarizedHeatmaps("o", "TCC", 45),
                        f = polarizedHeatmaps("f", "TCC", 56),
                        g = polarizedHeatmaps("g", "TCC", 67),
                        s = polarizedHeatmaps("s", "TCC", 78))

ringplot.tcc.ls
```

```{r}
lapply(as.character(all.taxlevs), function(x) ggsave(plot = ringplot.tcc.ls[[x]],
                                                     filename = 
                                                       paste0("../figures/",
                                                              "heatmap_ring_TCC_",
                                                              x, ".pdf"),
                                                     dev = "pdf"))
```

```{r}
ringplot.tcga.ls <- list(d = polarizedHeatmaps("d", "TCGA", 1),
                        k = polarizedHeatmaps("k", "TCGA", 12),
                        p = polarizedHeatmaps("p", "TCGA", 23),
                        c = polarizedHeatmaps("c", "TCGA", 34),
                        o = polarizedHeatmaps("o", "TCGA", 45),
                        f = polarizedHeatmaps("f", "TCGA", 56),
                        g = polarizedHeatmaps("g", "TCGA", 67),
                        s = polarizedHeatmaps("s", "TCGA", 78))

ringplot.tcga.ls
```

```{r}
lapply(as.character(all.taxlevs), function(x) ggsave(plot = ringplot.tcga.ls[[x]],
                                                     filename = 
                                                       paste0("../figures/",
                                                              "heatmap_ring_TCGA_",
                                                              x, ".pdf"),
                                                     dev = "pdf"))
```

## repeat for agreement plot

```{r}
agree.input <- surv.res %>%
  filter(cancer %in% good.cancers) %>%
  mutate(hazard.dir = ifelse(hazard.ratio > 1, "neg", "pos"),
         signif = p.value < 0.05,
         dirsig = paste(hazard.dir, signif)) %>%
  select(microbe, dirsig, cancer, datset) %>%
  spread(key = "datset", value = "dirsig") %>%
  separate(TCC, into = c("TCC.dir", "TCC.sig")) %>%
  separate(TCGA, into = c("TCGA.dir", "TCGA.sig")) %>%
  mutate(agree.code = TCC.sig == T & TCGA.sig == T & TCC.dir == TCGA.dir) %>%
  mutate(taxalev = str_sub(microbe,1,1))

saveRDS(agree.input, "../data/survival_concordance-table.RDS")
```

```{r}
agreeingHeatmaps <- function(tlev, yadj){
    agree.input %>%
    filter(taxalev == tlev) %>%
    filter(cancer %in% good.cancers) %>%
    mutate(cancer.code = ifelse(agree.code == T, cancer, NA),
           cancer = as.numeric(as.factor(cancer)) + yadj) %>%
    ggplot(aes(x = microbe, y = cancer, fill = cancer.code)) +
    geom_tile() +
    ylim(c(0,100)) +
    scale_fill_manual(values = vir.cols, name = "", na.value = "grey95") +
    theme_void() +
    theme(axis.text.x = element_blank(),
          axis.ticks = element_blank()) +
    coord_polar()
}
```

```{r}
ringplot.agree.ls <- list(d = agreeingHeatmaps("d", 1),
                          k = agreeingHeatmaps("k", 12),
                          p = agreeingHeatmaps("p", 23),
                          c = agreeingHeatmaps("c", 34),
                          o = agreeingHeatmaps("o", 45),
                          f = agreeingHeatmaps("f", 56),
                          g = agreeingHeatmaps("g", 67),
                          s = agreeingHeatmaps("s", 78))

ringplot.agree.ls
```

```{r}
lapply(as.character(all.taxlevs), function(x) ggsave(plot = ringplot.agree.ls[[x]],
                                                     filename = 
                                                       paste0("../figures/",
                                                              "heatmap_ring_agree_",
                                                              x, ".pdf"),
                                                     dev = "pdf"))
```

# Example curves

```{r}
poss.choices <- agree.input %>%
  filter(agree.code == T) %>%
  select(microbe, cancer) %>%
  left_join(surv.res) %>%
  filter(prevalence.1 > 10) %>%
  add_count(microbe, cancer) %>%
  filter(n>1)

```

```{r as forest plot}
poss.choices %>%
  mutate(microbe = gsub("s__(.*)", paste0("<i>", "\\1", "</i>"), microbe),
         microbe = gsub(" ", "<br>", microbe),
         microbe = gsub("sp.<br>", "sp.", microbe)) %>%
  ggplot(aes(x = microbe, y = log(hazard.ratio), fill = datset, group = datset,
             ymin = log(confint.low), ymax = log(confint.high))) +
  geom_col(position = "dodge") +
  geom_point(shape = "diamond", show.legend = F) +
  geom_errorbar(position = "dodge") +
  labs(x = "", y = "log(Hazard Ratio)") +
  scale_fill_manual(breaks = c("TCC", "TCGA"),
                    values = c("red", "blue"),
                    labels = c("ORIEN", "TCGA"),
                    name = "") +
  coord_flip() +
  theme_bw() +
  theme(text = element_text(size = 10),
        axis.text.y = element_markdown())

ggsave("../figures/forest_example-microbes.pdf", height = 4, width = 4)
```


# Check Phylum labels

```{r}
agree.input %>%
  filter(taxalev == "p") %>%
  filter(cancer %in% good.cancers) %>%
  mutate(cancer.code = ifelse(agree.code == T, cancer, NA),
         cancer = as.numeric(as.factor(cancer)) + 23) %>%
  ggplot(aes(x = microbe, y = cancer, fill = cancer.code)) +
  geom_tile() +
  geom_text(aes(label = ifelse(agree.code == T, microbe, NA))) +
  ylim(c(0,100)) +
  scale_fill_manual(values = vir.cols, name = "", na.value = "grey95") +
  theme_void() +
  theme(axis.text.x = element_blank(),
        axis.ticks = element_blank()) +
  coord_polar()
```

```{r}
surv.res %>%
  filter(microbe == "p__Spirochaetes" & cancer == "LUAD")
```
